### 2020-07-02 keurfonluu/toughio#52 @rreinecke
JOSS Review: Tests
* How can a potential user run the tests? Is that described somewhere?
* Do you supply example files? (I don't use Though so I don't have anything to test your software package properly, this might happen to a potential user as well). If so, where do I find a description on how to use them with your tool.
