+++ setup.py
@@ -51,8 +51,6 @@ else:
     EXTRA_REQUIREMENTS = {"full": ["pyvista >= 0.23.1"]}
 REQUIREMENTS += ["numpy >= 1.13.0"]
 EXTRA_REQUIREMENTS["full"] += ["matplotlib", "netcdf4", "h5py", "scipy>=0.9"]
-EXTRA_REQUIREMENTS["dev"] = list(EXTRA_REQUIREMENTS["full"])
-EXTRA_REQUIREMENTS["dev"] += ["black", "docformatter", "invoke", "isort", "pytest", "pytest-cov"]
 
 
 PACKAGE_DATA = {
