### 2020-07-03 41afbaa
add contributing guidelines
