### 2020-07-03 9c90ecb
add contributing guidelines
