+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,23 @@
+---
+name: Feature request
+about: Request the addition of a new feature/functionality
+
+---
+
+**Description of the desired feature**
+
+<!--
+Please be as detailed as you can in your description. If possible, include an
+example of how you would like to use this feature (even better if it's a code
+example).
+-->
+
+
+
+**Are you willing to help implement and maintain this feature?** Yes/No
+
+<!--
+Every feature we add is code that we will have to maintain and keep updated.
+This takes a lot of effort. If you are willing to be involved in the project
+and help maintain your feature, it will make it easier for us to accept it.
+-->
