+++ README.md
@@ -1,151 +1,0 @@
-# toughio
-
-[![License](https://img.shields.io/badge/license-BSD--3--Clause-green)](https://github.com/keurfonluu/toughio/blob/master/LICENSE)
-[![Stars](https://img.shields.io/github/stars/keurfonluu/toughio?logo=github)](https://github.com/keurfonluu/toughio)
-[![Pyversions](https://img.shields.io/pypi/pyversions/toughio.svg?style=flat)](https://pypi.org/pypi/toughio/)
-[![Version](https://img.shields.io/pypi/v/toughio.svg?style=flat)](https://pypi.org/project/toughio)
-[![Downloads](https://pepy.tech/badge/toughio)](https://pepy.tech/project/toughio)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg?style=flat)](https://github.com/psf/black)
-[![Codacy Badge](https://img.shields.io/codacy/grade/037f16f4e7544afbb1b99bf0d66df4a6.svg?style=flat)](https://www.codacy.com/manual/keurfonluu/toughio?utm_source=github.com&utm_medium=referral&utm_content=keurfonluu/toughio&utm_campaign=Badge_Grade)
-[![Codecov](https://img.shields.io/codecov/c/github/keurfonluu/toughio.svg?style=flat)](https://codecov.io/gh/keurfonluu/toughio)
-[![Documentation Status](https://readthedocs.org/projects/toughio/badge/?version=latest)](https://toughio.readthedocs.io/en/latest/?badge=latest)
-
-[TOUGH](https://tough.lbl.gov/) (Transport Of Unsaturated Groundwater and Heat) is a general purpose numerical simulation software designed for fluid and heat flows of multiphase, multicomponent fluid mixtures in porous and fractured media developed at Lawrence Berkeley National Laboratory. It solves mass and energy balance equations that describe fluid and heat flow in multiphase and multicomponent systems. TOUGH handles all types of multiphase and multicomponent flow systems since the governing equations for fluid and heat flow have the same mathematical form. The nature and properties of fluid mixtures are described by thermophysical variables (e.g. density, viscosity, enthalpy) which are provided by an equation-of-state (EOS) module.
-
-**`toughio`** is an open-source library that provides tools to facilitate pre- and post-processing for TOUGH using the latest Python standards. It aims to make setting up of a TOUGH simulation user-friendly by relying on existing well-established Python packages:
-
--   [**`numpy`**](https://numpy.org/): vectorized calculation of N-dimensional arrays,
--   [**`meshio`**](https://github.com/nschloe/meshio): input/output for many mesh formats,
--   [**`pyvista`**](https://github.com/pyvista/pyvista): 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK).
-
-Note that the results of a TOUGH simulation are sensitive to the quality of the mesh (ideally, it should satisfy the orthogonality condition). A mesh that contains too many ill-shaped cells can potentially lead to unexpected results although the simulation converged successfully. **`toughio`** does not verify the quality of the mesh which is left to the discretion of the user.
-
-<p align="center">
-    <img alt="sample_co2" src="https://github.com/keurfonluu/toughio/blob/master/.github/sample.gif" width="50%"><br>
-    <em>Simulation of CO<sub>2</sub> leakage along a fault.<br>Mesh generated with Gmsh and animation exported by PyVista.</em>
-</p>
-
-## Features
-
-Meshing:
-
--   Create simple 2D cylindric or 2D/3D structured meshes similarly to TOUGH's built-in _MESHMAKER_,
[…]
