+++ README.rst
@@ -1,0 +1,200 @@
+toughio
+=======
+
+|License| |Stars| |Pyversions| |Version| |Downloads| |Code style: black| |Codacy Badge| |Codecov| |Documentation Status|
+
+`TOUGH <https://tough.lbl.gov/>`__ (Transport Of Unsaturated Groundwater and Heat) is a general purpose numerical simulation software designed for fluid and heat flows of multiphase, multicomponent fluid mixtures in porous and fractured media developed at Lawrence Berkeley National Laboratory. It solves mass and energy balance equations that describe fluid and heat flow in multiphase and multicomponent systems. TOUGH handles all types of multiphase and multicomponent flow systems since the governing equations for fluid and heat flow have the same mathematical form. The nature and properties of fluid mixtures are described by thermophysical variables (e.g. density, viscosity, enthalpy) which are provided by an equation-of-state (EOS) module.
+
+**toughio** is an open-source library that provides tools to
+facilitate pre- and post-processing for TOUGH using the latest Python
+standards. It aims to make setting up of a TOUGH simulation
+user-friendly by relying on existing well-established Python packages:
+
+-  `numpy <https://numpy.org/>`__: vectorized calculation of
+   N-dimensional arrays,
+-  `meshio <https://github.com/nschloe/meshio>`__: input/output
+   for many mesh formats,
+-  `pyvista <https://github.com/pyvista/pyvista>`__: 3D plotting
+   and mesh analysis through a streamlined interface for the
+   Visualization Toolkit (VTK).
+
+Note that the results of a TOUGH simulation are sensitive to the quality of the mesh (ideally, it should satisfy the orthogonality condition). A mesh that contains too many ill-shaped cells can potentially lead to unexpected results although the simulation converged successfully. **toughio** does not verify the quality of the mesh which is left to the discretion of the user.
+
+.. figure:: https://raw.githubusercontent.com/keurfonluu/toughio/master/.github/sample.gif
+   :alt: sample_co2
+   :figwidth: 50%
+   :align: center
+
+   Simulation of CO\ :sub:`2` leakage along a fault. Mesh generated with Gmsh and animation exported by PyVista.
+
+Features
+--------
+
+Meshing:
+
+-  Create simple 2D cylindric or 2D/3D structured meshes similarly to
+   TOUGH's built-in *MESHMAKER*,
+-  Import mesh generated by external softwares (e.g.
+   `Abaqus <https://www.3ds.com/products-services/simulia/products/abaqus/>`__,
+   `FLAC3D <https://www.itascacg.com/software/flac3d>`__,
+   `Gmsh <http://gmsh.info/>`__,
+   `LaGriT <https://meshing.lanl.gov/>`__),
+-  Export imported or generated mesh to a *MESH* file for TOUGH
+   assuming
+   `conformity <https://www.quora.com/What-is-non-conformal-mesh-in-CFD>`__
+   (and optionally write initial condition file *INCON*).
+
+Pre-processing:
+
+-  Easily add initial conditions, boundary conditions or other physical
+   properties (e.g. porosity, permeability) using the convenient class
+   :class:`toughio.Mesh`,
+-  Import previous TOUGH simulation input file into Python or define
+   simulation parameters in a human-readable and
+   `JSON <http://json.org/>`__\ able dictionary,
+-  Export simulation parameters dictionary to a TOUGH input file.
+
+Post-processing:
+
+-  Import outputs of a TOUGH simulation into Python,
+-  Visualize results directly in Python using
+   `pyvista <https://github.com/pyvista/pyvista>`__ or export
+   the results to another format supported by
+   `meshio <https://github.com/nschloe/meshio>`__ (e.g. VTK,
[…]
