### 2020-07-06 keurfonluu/toughio#52 @rreinecke
The first question has been addressed by #53
Second would be important though for me to able to verify if your tool does what it is suppose to do
