### 2020-07-07 b0b18cf
update doc requirements
