+++ doc/requirements.txt
@@ -1,4 +1,6 @@
+pygmsh
 sphinx-argparse
+sphinx-gallery
 sphinx-rtd-theme
 sphinxcontrib-bibtex
 -e git://github.com/Naeka/pybtex-apa-style.git#egg=pybtex-apa-style
