+++ doc/source/example.rst
@@ -1,401 +1,0 @@
-Sample problem
-==============
-
-This section intends to describe a simple yet complete workflow from setting up a TOUGH simulation with :mod:`toughio` to importing and visualizing simulation results in Python.
-
-The model geometry used is inspired from the paper:
-
-..
-
-    Cappa, Frédéric, and Jonny Rutqvist. "Modeling of Coupled Deformation and Permeability Evolution during Fault Reactivation Induced by Deep Underground Injection of CO2." International Journal of Greenhouse Gas Control 5, no. 2 (March 2011): 336–346. <https://doi.org/10.1016/j.ijggc.2010.08.005>.
-
-The fault modeled in this sample problem has already been reactivated by increasing the permeability from its original value.
-
-Note that only the main :mod:`toughio` features are introduced.
-
-
-Generate mesh in Python with Gmsh
----------------------------------
-
-This section describes how to generate a mesh using Gmsh in Python. It requires PyGmsh to be installed:
-
-.. code-block::
-
-    pip install pygmsh --user
-
-The mesh can also be generated from scratch directly in Gmsh either using its GUI and/or its internal scripting language.
-
-Note that :mod:`toughio` is not required at this preliminary stage of the pre-processing. This section only intends to show how the mesh used in this sample problem has been generated. The user is free to use any meshing software as long as the mesh format is supported by :mod:`toughio` (through :mod:`meshio`).
-
-First, we import numpy and pygmsh, and initialize a pygmsh.Geometry object that will be used to construct all the geometrical entities of the model.
-
-.. code-block:: python
-
-    import numpy
-    import pygmsh
-
-    geo = pygmsh.built_in.Geometry()
-
-We can define a bunch of useful variables such as the characteristic length or some parameters to characterize the model.
-
-.. code-block:: python
-
-    lc = 100.0                              # Characteristic length of mesh
-    xmin, xmax = 0.0, 2000.0                # X axis boundaries
-    zmin, zmax = -500.0, -2500.0            # Z axis boundaries
-
-    inj_z = -1500.0                         # Depth of injection
-    flt_offset = 50.0                       # Offset of fault
-    flt_thick = 10.0                        # Thickness of fault
-    tana = numpy.tan(numpy.deg2rad(80.0))   # Tangeant of dipping angle of fault
-    dist = 500.0 - 0.5 * flt_thick          # Distance from injection point (0.0, -1500.0) to left wall of fault
-
-    bnd_thick = 10.0                        # Thickness of boundary elements
-
-We start by defining the geometrical entity representing the fault. The fault is represented as a finite-thickness element that intersects all the layers of the model. To ensure conformity of the final mesh, each wall of the fault is represented by a segmented line where the positions of the nodes correspond to the intersections of the hanging (left) and foot (right) walls with the different layers. Note that the foot wall is inverted so that the fault entity forms a closed rectangular loop.
-
-.. code-block:: python
-
-    depths = [zmin, -1300.0, -1450.0, -1550.0, -1700.0, zmax]
-    fault_left = [[dist + (z - inj_z) / tana, z, 0.0] for z in depths]
-
[…]
