+++ doc/source/conf.py
@@ -43,7 +43,32 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinxarg.ext",
     "sphinxcontrib.bibtex",
+    "sphinx_gallery.gen_gallery",
 ]
+
+# Sphinx Gallery settings
+from sphinx_gallery.sorting import FileNameSortKey
+
+sphinx_gallery_conf = {
+    "examples_dirs": [
+        "../../examples/co2_leakage_along_a_fault/",
+    ],
+    "gallery_dirs": [
+        "examples",
+    ],
+    "filename_pattern": r"\.py",
+    "download_all_examples": False,
+    "within_subsection_order": FileNameSortKey,
+    "backreferences_dir": None,
+    "doc_module": "toughio",
+    "image_scrapers": (
+        "matplotlib",
+        "pyvista",
+    ),
+    "first_notebook_cell": (
+        "%matplotlib inline\n"
+    ),
+}
 
 # Napoleon settings
 napoleon_google_docstring = False
@@ -59,7 +84,7 @@ napoleon_include_special_with_doc = False
 # napoleon_use_rtype = True
 
 # Numfig settings
-numfig = True
+numfig = False
 numfig_format = {
     "figure": "Figure %s",
 }
