+++ doc/source/gallery/index.rst
@@ -1,0 +1,10 @@
+Gallery
+========
+
+This gallery contains different sample problems to demonstrate :mod:`toughio` pre- and post-processing capabilities.
+
+
+.. toctree::
+   :titlesonly:
+
+   ../examples/index
