+++ doc/source/api/index.rst
@@ -1,0 +1,14 @@
+API Reference
+=============
+
+.. toctree::
+   :titlesonly:
+   :maxdepth: 2
+
+   io
+   mesh
+   rpcap
+   cli
+
+
+* :ref:`genindex`
