+++ doc/requirements.txt
@@ -1,4 +1,5 @@
 pygmsh
+seaborn
 sphinx-argparse
 sphinx-gallery
 sphinx-rtd-theme
