### 2020-07-07 ab9bccf
fix gallery_dirs
