+++ doc/source/gallery/index.rst
@@ -7,4 +7,4 @@ This gallery contains different sample problems to demonstrate :mod:`toughio` pr
 .. toctree::
    :titlesonly:
 
-   ../examples/index
+   ../examples/co2_leakage_along_a_fault/index
