### 2020-07-07 89ff92f
fix gallery_dirs
