+++ examples/co2_leakage_along_a_fault/2_generate_mesh_and_incon_files.py
@@ -85,7 +85,7 @@ mesh.add_cell_data("initial_condition", incon)
 ########################################################################################
 
 ########################################################################################
-# :mod:`toughio` also recognizes the cell data keys `"porosity"` and `"permeability"` in case we want to initialize porosity and/or permeability fields (e.g. if well logs data are available). Like boundary and initial conditions, we only have to associate new cell data arrays to keys `"porosity"` and/or `"permeability"`. The way these arrays are generated does not matter, they can be the results of simple interpolations (e.g. with :mod":`scipy`) or more advanced geostatistical interpolations (e.g. with :mod:`pykrige`).
+# :mod:`toughio` also recognizes the cell data keys `"porosity"` and `"permeability"` in case we want to initialize porosity and/or permeability fields (e.g. if well logs data are available). Like boundary and initial conditions, we only have to associate new cell data arrays to keys `"porosity"` and/or `"permeability"`. The way these arrays are generated does not matter, they can be the results of simple interpolations (e.g. with :mod:`scipy`) or more advanced geostatistical interpolations (e.g. with :mod:`pykrige`).
 
 ########################################################################################
 
