### 2020-07-07 2e63ebe
output coordinates
