+++ .codacy.yml
@@ -8,3 +8,4 @@ exclude_paths:
   - 'doc/source/conf.py'
   - 'tasks.py'
   - '.github/**'
+  - 'examples/**'
