### 2020-07-07 4978625
more git-lfs hack for rtd
