### 2020-07-07 e4fd481
Update docs after building Travis build 9 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
6acca8f98a4aee2ea7e972b628a7ca16b5fc4f47.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358618614.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy doc --built-docs ./doc/build/
