+++ doc/_sources/examples/co2_leakage_along_a_fault/2_generate_mesh_and_incon_files.rst.txt
@@ -1,0 +1,219 @@
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        Click :ref:`here <sphx_glr_download_examples_co2_leakage_along_a_fault_2_generate_mesh_and_incon_files.py>`     to download the full example code
+    .. rst-class:: sphx-glr-example-title
+
+    .. _sphx_glr_examples_co2_leakage_along_a_fault_2_generate_mesh_and_incon_files.py:
+
+
+Generate MESH and INCON files
+=============================
+
+In this example, we assume that the mesh has already been generated by a third-party software (here Gmsh via :mod:`pygmsh`).
+
+First, we import :mod:`numpy` and :mod:`toughio`.
+
+
+.. code-block:: default
+
+
+    import numpy
+    import toughio
+
+
+
+
+
+
+
+
+A supported mesh can be read using the function :func:`toughio.read_mesh` that returns a :class:`toughio.Mesh` object.
+
+
+.. code-block:: default
+
+
+    mesh = toughio.read_mesh("mesh.msh")
+
+
+
+
+
+
+
+
+The mesh used in this sample problem is 2D and has been defined in the XY plane, but the points have 3D coordinates (with zeros as 3rd dimension for every cells). To make it 3D in the XZ plane, we swap the 2nd and 3rd dimensions, and then extrude the mesh by 1 meter along the Y axis (2nd dimension).
+
+
+.. code-block:: default
+
+
+    mesh.points[:, [1, 2]] = mesh.points[:, [2, 1]]
+    mesh.extrude_to_3d(height=1.0, axis=1)
+
+
+
+
+
+
+
+
+(Optional) Before going any further, it is good practice to first check the quality of the mesh generated.
+TOUGH does not use any geometrical coordinate system and assumes that the line connecting a cell with its neighbor is orthogonal to their common interface. :mod:`toughio` provides a mesh property that measures the quality of a cell as the average absolute cosine angle between the line connecting a cell with its neighbor and the normal vector of the common interface.
+The mesh used in this example is of rather good quality. Bad quality cells are located at the boundaries of the model and mostly belong to the material `"BOUND"`. As this material is only used to impose Dirichlet boundary conditions in TOUGH, these bad quality cells will not impact the simulation outputs.
+
+
+.. code-block:: default
+
+
+    import pyvista
+    pyvista.set_plot_theme("document")
+
+    p = pyvista.Plotter(window_size=(800, 800))
+    p.add_mesh(
+        mesh=mesh.to_pyvista(),
+        scalars=mesh.qualities,
+        stitle="Average cell quality",
+        clim=(0.0, 1.0),
+        cmap="RdBu",
+        show_scalar_bar=True,
+        show_edges=True,
+    )
+    p.view_xz()
+    p.show()
+
+
+
+
+.. image:: /examples/co2_leakage_along_a_fault/images/sphx_glr_2_generate_mesh_and_incon_files_001.png
+    :alt: 2 generate mesh and incon files
+    :class: sphx-glr-single-img
+
+
+.. rst-class:: sphx-glr-script-out
+
+ Out:
+
+ .. code-block:: none
+
+
+    [(1005.0, -5504.599664394268, -1500.0),
+     (1005.0, 0.5, -1500.0),
+     (0.0, 0.0, 1.0)]
+
+
+
+(Optional) Usually, a simple distribution plot is enough to rapidly assess the quality of a mesh.
+
+
+.. code-block:: default
+
+
+    import seaborn
+
+    ax = seaborn.distplot(mesh.qualities[mesh.materials != "BOUND"], kde=False)
+
+
+
+
[…]
