+++ doc/_sources/examples/co2_leakage_along_a_fault/sg_execution_times.rst.txt
@@ -1,0 +1,18 @@
+
+:orphan:
+
+.. _sphx_glr_examples_co2_leakage_along_a_fault_sg_execution_times:
+
+Computation times
+=================
+**00:09.153** total execution time for **examples_co2_leakage_along_a_fault** files:
+
++--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_examples_co2_leakage_along_a_fault_2_generate_mesh_and_incon_files.py` (``2_generate_mesh_and_incon_files.py``)                                           | 00:06.268 | 0.0 MB |
++--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_examples_co2_leakage_along_a_fault_4_import_and_visualize_simulation_outputs_in_pyvista.py` (``4_import_and_visualize_simulation_outputs_in_pyvista.py``) | 00:01.771 | 0.0 MB |
++--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_examples_co2_leakage_along_a_fault_1_generate_mesh_in_python_with_gmsh.py` (``1_generate_mesh_in_python_with_gmsh.py``)                                   | 00:01.088 | 0.0 MB |
++--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_examples_co2_leakage_along_a_fault_3_generate_model_parameters_input_file.py` (``3_generate_model_parameters_input_file.py``)                             | 00:00.026 | 0.0 MB |
++--------------------------------------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
