+++ doc/_sources/index.rst.txt
@@ -1,0 +1,20 @@
+toughio's documentation
+=======================
+
+.. include:: ../../README.rst
+
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+   :caption: Getting started
+
+   gallery/index
+
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+   :caption: Reference documentation
+   
+   api/index

[…]
