### 2020-07-07 e172bf5
Update docs after building Travis build 12 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
43404311ddda85b2c5e6e9441d0c9db844002ba4.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358622795.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
