+++ _downloads/63e6fc27943783755535e4c8d94b8bc3/1_generate_mesh_in_python_with_gmsh.py
@@ -1,0 +1,323 @@
+"""
+Generate mesh in Python with Gmsh
+=================================
+
+This examples describes how to generate a mesh using Gmsh in Python. It requires :mod:`pygmsh` to be installed:
+
+.. code-block:: bash
+
+    pip install pygmsh --user
+
+The mesh can also be generated from scratch directly in Gmsh either using its GUI and/or its internal scripting language.
+
+Note that :mod:`toughio` is not required at this preliminary stage of the pre-processing. This example only intends to show how the mesh used in this sample problem has been generated. The user is free to use any meshing software as long as the mesh format is supported by :mod:`toughio` (through :mod:`meshio`).
+
+"""
+
+########################################################################################
+# First, we import :mod:`numpy` and :mod:`pygmsh`, and initialize a :class:`pygmsh.Geometry` object that will be used to construct all the geometrical entities of the model.
+
+import numpy
+import pygmsh
+
+geo = pygmsh.built_in.Geometry()
+
+########################################################################################
+
+########################################################################################
+# We can define a bunch of useful variables such as the characteristic length or some parameters to characterize the model.
+
+lc = 100.0                              # Characteristic length of mesh
+xmin, xmax = 0.0, 2000.0                # X axis boundaries
+zmin, zmax = -500.0, -2500.0            # Z axis boundaries
+
+inj_z = -1500.0                         # Depth of injection
+flt_offset = 50.0                       # Offset of fault
+flt_thick = 10.0                        # Thickness of fault
+tana = numpy.tan(numpy.deg2rad(80.0))   # Tangeant of dipping angle of fault
+dist = 500.0 - 0.5 * flt_thick          # Distance from injection point (0.0, -1500.0) to left wall of fault
+
+bnd_thick = 10.0                        # Thickness of boundary elements
+
+########################################################################################
+
+########################################################################################
+# We start by defining the geometrical entity representing the fault.
+# The fault is represented as a finite-thickness element that intersects all the layers of the model.
+# To ensure conformity of the final mesh, each wall of the fault is represented by a segmented line where the positions of the nodes correspond to the intersections of the hanging (left) and foot (right) walls with the different layers.
+# Note that the foot wall is inverted so that the fault entity forms a closed rectangular loop.
+
+depths = [zmin, -1300.0, -1450.0, -1550.0, -1700.0, zmax]
+fault_left = [[dist + (z - inj_z) / tana, z, 0.0] for z in depths]
+
+depths = [zmin, -1300.0 + flt_offset, -1450.0 + flt_offset, -1550.0 + flt_offset, -1700.0 + flt_offset, zmax]
+fault_right = [[dist + (z - inj_z) / tana + flt_thick, z, 0.0] for z in depths]
+
+fault = geo.add_polygon(
+    X=fault_left + fault_right[::-1],
+    lcar=0.1 * lc,
+)
+
+########################################################################################
+
+########################################################################################
[…]
