+++ _sources/examples/co2_leakage_along_a_fault/3_generate_model_parameters_input_file.rst.txt
@@ -1,0 +1,285 @@
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        Click :ref:`here <sphx_glr_download_examples_co2_leakage_along_a_fault_3_generate_model_parameters_input_file.py>`     to download the full example code
+    .. rst-class:: sphx-glr-example-title
+
+    .. _sphx_glr_examples_co2_leakage_along_a_fault_3_generate_model_parameters_input_file.py:
+
+
+Generate model parameters input file
+====================================
+
+Now that the mesh has been pre-processed, we can define the TOUGH simulation parameters.
+
+As always, we first import the required modules for this example.
+
+
+.. code-block:: default
+
+
+    import numpy
+    import toughio
+
+
+
+
+
+
+
+
+A :mod:`toughio` input file is defined as a nested dictionary with meaningful keywords.
+Let's initialize our parameter dictionary by giving the simulation a title and defining the equation-of-state. This example will also be run isothermally.
+
+
+.. code-block:: default
+
+
+    parameters = {
+        "title": "Simulation of CO2 leak along a fault",
+        "eos": "eco2n",
+        "isothermal": True,
+    }
+
+
+
+
+
+
+
+
+We can also define some default values that are shared by the different materials.
+
+
+.. code-block:: default
+
+
+    parameters["default"] = {
+        "density": 2260.0,
+        "conductivity": 1.8,
+        "specific_heat": 1500.0,
+        "compressibility": 8.33e-10,
+        "conductivity_dry": 1.8,
+        "tortuosity": 0.7,
+        "relative_permeability": {
+            "id": 3,
+            "parameters": [0.3, 0.05],
+        },
+        "capillarity": {
+            "id": 7,
+            "parameters": [0.457, 0.0, 5.03e-5, 5.0e7, 0.99],
+        },
+    }
+
+
+
+
+
+
+
+
+Now, we define specific material properties (different than the default ones previously defined) for each material in the mesh (i.e. we write the block `ROCKS`).
+
+
+.. code-block:: default
+
+
+    parameters["rocks"] = {
+        "UPPAQ": {
+            "porosity": 0.10,
+            "permeability": 1.0e-14,
+        },
+        "CENAQ": {
+            "porosity": 0.10,
+            "permeability": 1.0e-13,
+        },
+        "BASAQ": {
+            "porosity": 0.01,
+            "permeability": 1.0e-16,
+            "relative_permeability": {
+                "id": 3,
+                "parameters": [0.3, 0.05],
+            },
+            "capillarity": {
+                "id": 7,
+                "parameters": [0.457, 0.0, 1.61e-6, 5.0e7, 0.99],
+            },
+        },
+        "CAPRO": {
+            "porosity": 0.01,
+            "permeability": 1.0e-19,
+            "relative_permeability": {
+                "id": 3,
+                "parameters": [0.3, 0.05],
+            },
+            "capillarity": {
+                "id": 7,
+                "parameters": [0.457, 0.0, 1.61e-6, 5.0e7, 0.99],
+            },
+        },
+        "FAULT": {
+            "porosity": 0.10,
+            "permeability": 1.0e-13,
+        },
+        "BOUND": {
+            "specific_heat": 1.0e55,
+            "porosity": 0.10,
+            "permeability": 1.0e-13,
+        },
+    }
+
+
+
+
+
+
+
+
+We can specify some simulation parameters (block `PARAM`), options (`MOP`) and selections (block `SELEC`).
+
+
+.. code-block:: default
+
+
[…]

[…]
