+++ README.rst
@@ -84,7 +84,7 @@ location:
 Documentation
 -------------
 
-Refer to the online `documentation <https://toughio.readthedocs.io/en/latest/>`__ for detailed description of the API and examples.
+Refer to the online `documentation <https://keurfonluu.github.io/toughio//>`__ for detailed description of the API and examples.
 
 Alternatively, the documentation can be built using `Sphinx <https://www.sphinx-doc.org/en/master/>`__
 
