### 2020-07-07 4a10498
Update docs after building Travis build 13 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
9474a7f11cbadb603cd6deab004cf90c937e9d02.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358626675.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
