### 2020-07-07 keurfonluu/toughio#52 @keurfonluu
Hi,

I've added a complete example problem (from pre- to post-processing) in the documentation:
https://keurfonluu.github.io/toughio//examples/co2_leakage_along_a_fault/index.html

This gallery is automatically generated from [`./examples`](https://github.com/keurfonluu/toughio/tree/master/examples). Input and output files are included so all scripts should run (even without Gmsh and TOUGH).

Note that the documentation has been moved from ReadtheDocs to GitHub-Pages since RTD is too limited in terms of functionalities (no support for git-lfs and headless display in my case).

Otherwise, other minor features can be automatically tested running `pytest` (91% coverage right now, the remaining 9% mostly corresponding to input argument checking).

For both doc and test, some files are stored as LFS, please run `git lfs pull` before running the example scripts and automated tests.
