### 2020-07-08 keurfonluu/toughio#52 @rreinecke
Great @keurfonluu ! Could you please add these important hints also to the README. This will help potential users to easily explore everything your project has to offer.
