### 2020-07-08 keurfonluu/toughio#52 @rreinecke
Downloading examples/co2_leakage_along_a_fault/CO2TAB (624 KB)
Error downloading object: examples/co2_leakage_along_a_fault/CO2TAB (d5376c1): Smudge error: Error downloading examples/co2_leakage_along_a_fault/CO2TAB (d5376c1145adba382418b4fa604074a10aaa99162e49602e789a92b75138da14): batch response: This repository is over its data quota. Account responsible for LFS bandwidth should purchase more data packs to restore access.

Errors logged to /tmp/toughio/.git/lfs/logs/20200708T071124.214786408.log
Use `git lfs logs last` to view the log.
error: external filter 'git-lfs filter-process' failed
fatal: examples/co2_leakage_along_a_fault/CO2TAB: smudge filter lfs failed
