### 2020-07-08 keurfonluu/toughio#52 @keurfonluu
Yeah, it seems that Travis CI has consumed my monthly LFS bandwidth quota in a matter of hours (which seems to be a common [issue](https://github.com/travis-ci/travis-ci/issues/8787). I'll investigate and then restore access to LFS. Otherwise, I'll have to give up on LFS.
