+++ README.rst
@@ -5,18 +5,11 @@ toughio
 
 `TOUGH <https://tough.lbl.gov/>`__ (Transport Of Unsaturated Groundwater and Heat) is a general purpose numerical simulation software designed for fluid and heat flows of multiphase, multicomponent fluid mixtures in porous and fractured media developed at Lawrence Berkeley National Laboratory. It solves mass and energy balance equations that describe fluid and heat flow in multiphase and multicomponent systems. TOUGH handles all types of multiphase and multicomponent flow systems since the governing equations for fluid and heat flow have the same mathematical form. The nature and properties of fluid mixtures are described by thermophysical variables (e.g. density, viscosity, enthalpy) which are provided by an equation-of-state (EOS) module.
 
-**toughio** is an open-source library that provides tools to
-facilitate pre- and post-processing for TOUGH using the latest Python
-standards. It aims to make setting up of a TOUGH simulation
-user-friendly by relying on existing well-established Python packages:
+**toughio** is an open-source library that provides tools to facilitate pre- and post-processing for TOUGH using the latest Python standards. It aims to make setting up of a TOUGH simulation user-friendly by relying on existing well-established Python packages:
 
--  `numpy <https://numpy.org/>`__: vectorized calculation of
-   N-dimensional arrays,
--  `meshio <https://github.com/nschloe/meshio>`__: input/output
-   for many mesh formats,
--  `pyvista <https://github.com/pyvista/pyvista>`__: 3D plotting
-   and mesh analysis through a streamlined interface for the
-   Visualization Toolkit (VTK).
+-  `numpy <https://numpy.org/>`__: vectorized calculation of N-dimensional arrays,
+-  `meshio <https://github.com/nschloe/meshio>`__: input/output for many mesh formats,
+-  `pyvista <https://github.com/pyvista/pyvista>`__: 3D plotting and mesh analysis through a streamlined interface for the Visualization Toolkit (VTK).
 
 Note that the results of a TOUGH simulation are sensitive to the quality of the mesh (ideally, it should satisfy the orthogonality condition). A mesh that contains too many ill-shaped cells can potentially lead to unexpected results although the simulation converged successfully. **toughio** does not verify the quality of the mesh which is left to the discretion of the user.
 
@@ -32,54 +25,43 @@ Features
 
 Meshing:
 
--  Create simple 2D cylindric or 2D/3D structured meshes similarly to
-   TOUGH's built-in *MESHMAKER*,
--  Import mesh generated by external softwares (e.g.
-   `Abaqus <https://www.3ds.com/products-services/simulia/products/abaqus/>`__,
-   `FLAC3D <https://www.itascacg.com/software/flac3d>`__,
-   `Gmsh <http://gmsh.info/>`__,
-   `LaGriT <https://meshing.lanl.gov/>`__),
--  Export imported or generated mesh to a *MESH* file for TOUGH
-   assuming
-   `conformity <https://www.quora.com/What-is-non-conformal-mesh-in-CFD>`__
-   (and optionally write initial condition file *INCON*).
+-  Create simple 2D cylindric or 2D/3D structured meshes similarly to TOUGH's built-in *MESHMAKER*,
+-  Import mesh generated by external softwares (e.g. `Abaqus <https://www.3ds.com/products-services/simulia/products/abaqus/>`__, `FLAC3D <https://www.itascacg.com/software/flac3d>`__, `Gmsh <http://gmsh.info/>`__, `LaGriT <https://meshing.lanl.gov/>`__),
+-  Export imported or generated mesh to a *MESH* file for TOUGH assuming `conformity <https://www.quora.com/What-is-non-conformal-mesh-in-CFD>`__ (and optionally write initial condition file *INCON*).
 
 Pre-processing:
 
[…]
