### 2020-07-08 7cc5d72
add badge docs from travis
