### 2020-07-08 bfa1ae3
Update docs after building Travis build 18 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
ac106985b44f82ef48f66bb153011e43c8c354cf.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358720291.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
