### 2020-07-08 keurfonluu/toughio#52 @keurfonluu
I found a workaround and reenabled access to LFS by purchasing a data pack. You should be good to go.

> Great @keurfonluu ! Could you please add these important hints also to the README. This will help potential users to easily explore everything your project has to offer.

Done ! Thanks !
