### 2020-07-08 e99e9ca
Update docs after building Travis build 19 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
9c757d56010cd12e67b5559be6e975bc3dcae9a6.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358735788.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
