+++ _sources/examples/co2_leakage_along_a_fault/4_import_and_visualize_simulation_outputs_in_pyvista.rst.txt
@@ -109,7 +109,7 @@ Once output data have been imported, we can simply call the method :meth:`toughi
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.772 seconds)
+   **Total running time of the script:** ( 0 minutes  1.798 seconds)
 
 
 .. _sphx_glr_download_examples_co2_leakage_along_a_fault_4_import_and_visualize_simulation_outputs_in_pyvista.py:
