### 2020-07-08 1264a47
Update docs after building Travis build 20 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
a38e9b5ed3c7e184908944b2467842b71b73dd3b.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358736979.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
