### 2020-07-08 keurfonluu/toughio#52 @rreinecke
> Otherwise, other minor features can be automatically tested running `pytest` (91% coverage right now, the remaining 9% mostly corresponding to input argument checking).

pytest only works if the module is already installed. With a clean clone python is not able to find the module.
