### 2020-07-08 b6a8c0e
Update docs after building Travis build 21 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
fa188c29fa0efa1ddfd09201776b018df3085ac5.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/358745431.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
