### 2020-07-08 keurfonluu/toughio#52 @keurfonluu
> pytest only works if the module is already installed. With a clean clone python is not able to find the module.

You are right, it should be clearer in the README now.
