### 2020-07-08 keurfonluu/toughio#52 @rreinecke
> > pytest only works if the module is already installed. With a clean clone python is not able to find the module.
>
> You are right, it should be clearer in the README now.

It should be possible to also load the local modules with pytest without having to install everything first. But thats also a solution that works for me
