### 2020-07-08 #2412 @rreinecke
> A very clean and well prepared submission @keurfonluu
> Only a very small issue with the test currently open in [keurfonluu/toughio#52](https://github.com/keurfonluu/toughio/issues/52)
> Besides that I think it is good to go from my side.

Has been fixed. All good to go.
