### 2020-07-10 8c8de12
add default material name option
