### 2020-07-13 keurfonluu/toughio#59 @ma-sadeghi
Automated tests missing
I couldn't find a reference to automated testing in your package. If that's the case, you can easily set it up using Gtihub Actions or other services like Travis-CI.
