### 2020-07-13 keurfonluu/toughio#60 @ma-sadeghi
Mention a reference to your issue tracker
It would be nice to mention a reference to your issue tracker (so users know where to ask questions/report bugs) both in your Github homepage as well as in the documentation.
