### 2020-07-13 #2412 @ma-sadeghi
- [ ] Documentation (keurfonluu/toughio#60)
- [ ] Automated testing (keurfonluu/toughio#59)
