### 2020-07-13 keurfonluu/toughio#59 @keurfonluu
Hi,

As described in the README, you can run `pytest` locally to automatically test the package. Automatic tests are run on pull request on GitHub Actions, and documentation is built and deployed on Travis CI.
