### 2020-07-13 keurfonluu/toughio#59 @ma-sadeghi
Sorry, I didn't know you already set up Github Actions.

A suggestion: I think it would be nice to put the CI badge in your README file. I already found the url for your repo:
https://github.com/keurfonluu/toughio/workflows/Python%20package/badge.svg
