+++ README.rst
@@ -146,7 +146,7 @@ Contributing
 ------------
 
 Please refer to the `Contributing
-Guidelines <https://github.com/keurfonluu/toughio/blob/master/CONTRIBUTING.md>`__ to see how you can help. This project is released with a `Code of Conduct <https://github.com/keurfonluu/toughio/blob/master/CODE_OF_CONDUCT.md>`__ which you agree to abide by when contributing.
+Guidelines <https://github.com/keurfonluu/toughio/blob/master/CONTRIBUTING.rst>`__ to see how you can help. This project is released with a `Code of Conduct <https://github.com/keurfonluu/toughio/blob/master/CODE_OF_CONDUCT.rst>`__ which you agree to abide by when contributing.
 
 Notice
 ------
