### 2020-07-13 021a292
Update docs after building Travis build 23 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
ee8e615f6d6449c7e02ce3eeb6331280c3912ca8.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/360606484.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
