### 2020-07-13 253cd3b
Update docs after building Travis build 24 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
5957f68d271a4155fe6f287bc54845585b62e41f.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/360607159.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
