+++ _sources/examples/co2_leakage_along_a_fault/3_generate_model_parameters_input_file.rst.txt
@@ -254,7 +254,7 @@ At this stage, all the input files required to run the simulation have been gene
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.055 seconds)
+   **Total running time of the script:** ( 0 minutes  0.028 seconds)
 
 
 .. _sphx_glr_download_examples_co2_leakage_along_a_fault_3_generate_model_parameters_input_file.py:
