### 2020-07-13 keurfonluu/toughio#60 @keurfonluu
Both the README and documentation are referring to these [Contributing Guidelines](https://github.com/keurfonluu/toughio/blob/master/CONTRIBUTING.rst).
