### 2020-07-13 keurfonluu/toughio#59 @keurfonluu
I've added a build badge (customized with shields.io to be consistent with the other badges, though).
