### 2020-07-13 #2412 @jgostick
Hi @ma-sadeghi Automated testing is not always necessary, since someone can write solid scientific code that does not know how to automate their testing.  The only requirement is that there is some sort of testing, like a script with several cases in it...more like un-automated integration tests.  Given this, do you still think they need to address the testing part?
