### 2020-07-13 #2412 @ma-sadeghi
@jgostick I was just following the JOSS checklist (in which "Automated tests" is listed)

Anyways, it turns out they do have automated testing, the badge was just not showing up, which is now fixed. It's all good now.
