+++ toughio/_common.py
@@ -37,6 +37,7 @@ block_to_format = {
     },
     "DIFFU": ",".join(8 * ["10.3e"]),
     "OUTPU": {1: "20s", 2: "15s", 3: "20s,5d,5d"},
+    "COORD": ",".join(3 * ["20.13e"]),
     "ELEME": {
         5: "5s,5d,5d,5s,10.4e,10.4e,10.4e,10.3e,10.3e,10.3e",
         6: "6s,5d,4d,5s,10.4e,10.4e,10.4e,10.3e,10.3e,10.3e",
