+++ toughio/_mesh/tough/_helpers.py
@@ -48,6 +48,19 @@ def _write_eleme(labels, materials, volumes, nodes, material_name=None):
         yield record
 
 
+def _write_coord(nodes):
+    """Return a generator that iterates over the records of block COORD."""
+    fmt = "{:20.13e}{:20.13e}{:20.13e}\n"
+    
+    for node in nodes:
+        record = fmt.format(
+            node[0],  # X
+            node[1],  # Y
+            node[2],  # Z
+        )
+        yield record
+
+
 def _write_conne(clabels, isot, d1, d2, areas, angles):
     """Return a generator that iterates over the records of block CONNE."""
     label_length = len(clabels[0][0])
