### 2020-07-16 1a5be03
fix none in csv output
