### 2020-07-16 4b7f722
Update docs after building Travis build 29 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
9f6befb31e06ec5c9db55214170abcffb4232b50.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/361509969.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
