### 2020-07-16 #2412 @jgostick
@keurfonluu Everything looks great.  My only suggestion would be to add some sort of concluding remark(s) to the paper after the figure, since it seems to end rather abruptly.  This is only a suggestion though.

Here is a list of what to do next:
