### 2020-07-24 6a995f4
add short conclusion
