+++ paper/paper.md
@@ -34,6 +34,8 @@ Figure 1 shows the result of a sample CO<sub>2</sub> sequestration simulation wh
 
 ![Example of simulation of CO<sub>2</sub> upward leakage along a fault completely developed with `toughio`. Mesh has been generated with Gmsh and imported in Python by `meshio`. Output figure has been prepared and exported by `pyvista`.](./figures/sample.png)
 
+`toughio` offers a complete set of features to pre- and post-process a TOUGH simulation. Mesh generated by third-party softwares can be imported and converted to a TOUGH MESH. TOUGH input files can be generated and/or edited directly in Python. Simulation outputs can be post-processed and visualized in Python.
+
 # Acknowledgements
 
 This material is based upon work supported by Lawrence Berkeley National Laboratory under U.S. Department of Energy Award No. DE-AC02-05CH11231.
