+++ paper/paper.md
@@ -34,7 +34,7 @@ Figure 1 shows the result of a sample CO<sub>2</sub> sequestration simulation wh
 
 ![Example of simulation of CO<sub>2</sub> upward leakage along a fault completely developed with `toughio`. Mesh has been generated with Gmsh and imported in Python by `meshio`. Output figure has been prepared and exported by `pyvista`.](./figures/sample.png)
 
-`toughio` offers a complete set of features to pre- and post-process a TOUGH simulation. Mesh generated by third-party softwares can be imported and converted to a TOUGH MESH. TOUGH input files can be generated and/or edited directly in Python. Simulation outputs can be post-processed and visualized in Python.
+`toughio` offers a complete set of features to pre- and post-process a TOUGH simulation in Python. Finite-element meshes generated by third-party softwares can be imported and converted to a Voronoi graph for TOUGH, simulation parameters can be defined using a human-readable and jsonable dictionary automatically converted to a fixed-format input file for TOUGH, and simulation results can be imported in Python for post-processing and visualization.
 
 # Acknowledgements
 
