### 2020-07-27 401c847
install hdf5 and netcdf
