+++ .github/workflows/ci.yml
@@ -18,6 +18,10 @@ jobs:
       uses: actions/setup-python@v1
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Install system dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y libhdf5-dev libnetcdf-dev
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
