### 2020-07-27 b715bf5
Update docs after building Travis build 37 of keurfonluu/toughio
The docs were built from the branch 'master' against the commit
07b2078af1e22fb20c0dfd8395d743d7513b9f4f.

The Travis build that generated this commit is at
https://travis-ci.com/keurfonluu/toughio/jobs/365257618.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs ./doc/build/
