### 2020-07-27 #2412 @keurfonluu
Sorry for the delay, got a couple of deadlines.

I've added a short conclusion to the paper after the example. Please find below the release version number and DOI:
- version: v1.3.2
- DOI: 10.5281/zenodo.3961279
