### 2020-07-27 #2412 @kthyng
@keurfonluu I'm wrapping up the publication of your paper this week. Can you change the metadata on your Zenodo archive so that the title and author list exactly matches your paper? This is not a requirement, but it looks better and is preferred.
