### 2020-07-27 #2412 @kthyng
Congrats to @jgostick on your new publication!! Thanks so much for @rreinecke and @ma-sadeghi for reviewing, and to @jgostick for editing — this whole process cannot happen without you and your expertise!
